Obtain trophic level data from rfishbase

The trophic level will be obtained for each scientific name provided. Search will be in fishbase for vertibrates and sealifebase for invertibrates. Two fields are accessed from the ecology table, FoodTroph and DietTroph. See fishbase for more details (http://fishbase.us/manual/English/FishbaseThe_ECOLOGY_Table.htm)

get_trophic_level(lookupTable)

Arguments

lookupTable
dataframe. Any size. One column must be called SCIENTIFIC_NAME. Each row of the table will be cross referenced with fishbase

Value

Tibble. The same columns as lookupTable but with following additional columns.

genusSpecies

TRUE if species name is genus species, otherwise FALSE

DietTroph

Value from DietTroph field in fishbase ecology table

FoodTroph

Value from FoodTroph field in fishbase ecology table

vertibrate

TRUE if species found in fishbase, FALSE if found in sealifebase, NA if not found